+++ R/geom_textcontour.R
@@ -1,6 +1,6 @@
 ##---------------------------------------------------------------------------##
 ##                                                                           ##
-##  geom_labelcontour main                                                   ##
+##  geom_textcontour main                                                   ##
 ##                                                                           ##
 ##  Copyright (C) 2021 by Allan Cameron & Teun van den Brand                 ##
 ##                                                                           ##
@@ -15,7 +15,7 @@
 #'
 #' @description Contour lines are available already in \pkg{ggplot2}, but the
 #'   native [`geom_contour`][ggplot2::geom_contour] does not allow the lines to
-#'   be labelled with the level of each contour. `geom_labelcontour` adds this
+#'   be labelled with the level of each contour. `geom_textcontour` adds this
 #'   ability.
 #'
 #' @inheritParams ggplot2::layer
@@ -75,17 +75,19 @@
 #' @md
 #'
 #' @examples
-#' set.seed(1)
+#' df <- expand.grid(x = seq(nrow(volcano)), y = seq(ncol(volcano)))
+#' df$z <- as.vector(volcano)
 #'
-#' df  <- data.frame(x = rnorm(100), y = rnorm(100))
-#'
-#' ggplot(df, aes(x, y)) +
-#' stat_density2d(geom = "textpath", aes(label = after_stat(level))) +
-#'  theme_classic()
+#' ggplot(df, aes(x, y, z = z)) +
+#'   geom_contour_filled(bins = 6, alpha = 0.6) +
+#'   geom_textcontour(bins = 6, size = 2.5, padding = unit(0.05, "in")) +
+#'   scale_fill_manual(values = terrain.colors(11)) +
+#'   theme_classic() +
+#'   theme(legend.position = "none")
 
 
-geom_labelcontour <- function(
-  mapping = NULL, data = NULL, stat = "label_contour",
+geom_textcontour <- function(
+  mapping = NULL, data = NULL, stat = "text_contour",
   position = "identity", na.rm = FALSE, show.legend = NA,
   inherit.aes = TRUE,
   lineend = "butt", linejoin = "round", linemitre = 10,
@@ -95,7 +97,7 @@ geom_labelcontour <- function(
   ...
   )
 {
-  layer(geom = GeomLabelContour, mapping = mapping, data = data,
+  layer(geom = GeomTextContour, mapping = mapping, data = data,
         stat = stat,
         position = position, show.legend = show.legend,
         inherit.aes = inherit.aes,
@@ -116,12 +118,13 @@ geom_labelcontour <- function(
         ))
 }
 
-#' @rdname geom_labelcontour
+#' @rdname geom_textcontour
 #' @format NULL
 #' @usage NULL
 #' @export
 #' @include geom_textpath.R
-stat_labelcontour <- function(mapping = NULL, data = NULL,
+stat_textcontour <- function(mapping = NULL, data = NULL,
+                             geom = "text_contour",
                          position = "identity",
                          ...,
                          bins = NULL,
@@ -133,8 +136,8 @@ stat_labelcontour <- function(mapping = NULL, data = NULL,
   layer(
     data = data,
     mapping = mapping,
-    stat = StatLabelContour,
-    geom = GeomLabelContour,
+    stat = StatTextContour,
+    geom = geom,
     position = position,
     show.legend = show.legend,
     inherit.aes = inherit.aes,
@@ -148,12 +151,12 @@ stat_labelcontour <- function(mapping = NULL, data = NULL,
   )
 }
 
-#' @rdname geom_labelcontour
+#' @rdname geom_textcontour
 #' @format NULL
 #' @usage NULL
 #' @export
 #' @include geom_textpath.R
-GeomLabelContour <- ggproto("GeomLabelContour", GeomTextpath,
+GeomTextContour <- ggproto("GeomTextContour", GeomTextpath,
   required_aes = c("x", "y"),
   default_aes = aes(colour = "black", size = 3.88, hjust = 0.5, vjust = 0.5,
                     family = "", fontface = 1, lineheight = 1.2, alpha = 1,
@@ -161,11 +164,12 @@ GeomLabelContour <- ggproto("GeomLabelContour", GeomTextpath,
                     linecolour = "_copy_text_colour_", angle = 0)
 )
 
-#' @rdname geom_labelcontour
+#' @rdname geom_textcontour
 #' @format NULL
 #' @usage NULL
 #' @export
-StatLabelContour <- ggproto("StatLabelContour", StatContour,
+#' @include geom_textpath.R
+StatTextContour <- ggproto("StatTextContour", StatContour,
 
   required_aes = c("x", "y", "z"),
   default_aes = aes(order = after_stat(level),