+++ tests/testthat/test-geom_textcontour.R
@@ -3,8 +3,8 @@ test_that("We can produce labelled contour plots", {
   df   <- expand.grid(x = seq(nrow(volcano)), y = seq(ncol(volcano)))
   df$z <- as.vector(volcano)
 
-  expect_silent(ggplot(df, aes(x, y, z = z)) + geom_labelcontour())
-  expect_silent(ggplot(df, aes(x, y, z = z)) + stat_labelcontour())
+  expect_silent(ggplot(df, aes(x, y, z = z)) + geom_textcontour())
+  expect_silent(ggplot(df, aes(x, y, z = z)) + stat_textcontour())
 })
 
 test_that("We can produce labelled 2d density plots", {
@@ -13,5 +13,5 @@ test_that("We can produce labelled 2d density plots", {
 
   df <- data.frame(x = rnorm(100), y = rnorm(100))
 
-  expect_silent(ggplot(df, aes(x, y)) + geom_labeldensity2d())
+  expect_silent(ggplot(df, aes(x, y)) + geom_textdensity2d())
 })